perf: repository服务去掉对git命令依赖 #11193 迁移scm监控服务
